+++ app/Http/Controllers/FakultasController.php
@@ -4,11 +4,16 @@ namespace App\Http\Controllers;
 
 use App\Models\Fakultas;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class FakultasController extends Controller
 {
     public function index(Request $request)
     {
+        if (!Auth::user()->canCrud('fakultas')) {
+            abort(403, 'Unauthorized action.');
+        }
+
          $query = Fakultas::query();
 
         if ($search = $request->search) {
@@ -24,11 +29,20 @@ class FakultasController extends Controller
 
     public function create()
     {
+         if (!Auth::user()->canCrud('fakultas')) {
+            abort(403, 'Unauthorized action.');
+        }
+        
         return view('page.fakultas.create');
     }
 
     public function store(Request $request)
     {
+
+         if (!Auth::user()->canCrud('fakultas')) {
+            abort(403, 'Unauthorized action.');
+        }
+
         $request->validate([
             'nama_fakultas' => 'required|string|max:255',
             'dekan' => 'nullable|string|max:50',
@@ -41,12 +55,20 @@ class FakultasController extends Controller
 
     public function edit($id)
     {
+         if (!Auth::user()->canCrud('fakultas')) {
+            abort(403, 'Unauthorized action.');
+        }
+
         $fakultas = Fakultas::findOrFail($id);
         return view('page.fakultas.edit', compact('fakultas'));
     }
 
     public function update(Request $request, $id)
     {
+         if (!Auth::user()->canCrud('fakultas')) {
+            abort(403, 'Unauthorized action.');
+        }
+
         $request->validate([
             'nama_fakultas' => 'required|string|max:255',
             'dekan' => 'nullable|string|max:50',
@@ -61,6 +83,10 @@ class FakultasController extends Controller
 
     public function destroy($id)
     {
+         if (!Auth::user()->canCrud('fakultas')) {
+            abort(403, 'Unauthorized action.');
+        }
+
         $fakultas = Fakultas::findOrFail($id);
         $fakultas->delete();
         return redirect()->route('fakultas.index')->with('success', 'Data fakultas berhasil dihapus.');
