+++ app/Http/Controllers/ProdiController.php
@@ -5,11 +5,17 @@ namespace App\Http\Controllers;
 use App\Models\Prodi;
 use App\Models\Fakultas;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ProdiController extends Controller
 {
     public function index()
     {
+         if (!Auth::user()->canCrud('prodi')) {
+            abort(403, 'Unauthorized action.');
+        }
+
+
         $prodi = Prodi::with('fakultas')->latest()->paginate(15);
         $fakultas = Fakultas::all(); // <-- tambahkan ini
         return view('page.prodi.index', compact('prodi', 'fakultas'));
@@ -18,12 +24,20 @@ class ProdiController extends Controller
 
     public function create()
     {
+         if (!Auth::user()->canCrud('prodi')) {
+            abort(403, 'Unauthorized action.');
+        }
+
         $fakultas = Fakultas::all();
         return view('page.prodi.create', compact('fakultas'));
     }
 
     public function store(Request $request)
     {
+         if (!Auth::user()->canCrud('prodi')) {
+            abort(403, 'Unauthorized action.');
+        }
+
 
         $request->validate([
             'id_fakultas' => 'required|exists:fakultas,id',
@@ -38,6 +52,10 @@ class ProdiController extends Controller
 
     public function edit($id)
     {
+         if (!Auth::user()->canCrud('prodi')) {
+            abort(403, 'Unauthorized action.');
+        }
+
         $prodi = Prodi::findOrFail($id);
         $fakultas = Fakultas::all();
         return view('page.prodi.edit', compact('prodi', 'fakultas'));
@@ -45,6 +63,10 @@ class ProdiController extends Controller
 
     public function update(Request $request, $id)
     {
+         if (!Auth::user()->canCrud('prodi')) {
+            abort(403, 'Unauthorized action.');
+        }
+
         try {
             $request->validate([
                 'id_fakultas' => 'required|exists:fakultas,id', // PERBAIKAN: exists:fakultas,id
@@ -64,6 +86,10 @@ class ProdiController extends Controller
 
     public function destroy($id)
     {
+         if (!Auth::user()->canCrud('prodi')) {
+            abort(403, 'Unauthorized action.');
+        }
+        
         $prodi = Prodi::findOrFail($id);
         $prodi->delete();
         return redirect()->route('prodi.index')->with('success', 'Data program studi berhasil dihapus.');
